Cascading tiles on level refresh.

+++ LD37/Core/Timer.cs
@@ -10,12 +10,12 @@ namespace LD37.Core
 
 		private bool repeating;
 
-		public Timer(int duration, Action trigger, bool repeating, float initialElapsed = 0) :
+		public Timer(float duration, Action trigger, bool repeating, float initialElapsed = 0) :
 			this(duration, null, trigger, repeating)
 		{
 		}
 
-		public Timer(int duration, Action<float> tick, Action trigger, bool repeating, float initialElapsed = 0)
+		public Timer(float duration, Action<float> tick, Action trigger, bool repeating, float initialElapsed = 0)
 		{
 			this.tick = tick;
 			this.trigger = trigger;
